server-code/requirements.txt
absl-py==0.12.0
aiohttp==3.7.4.post0
astunparse==1.6.3
async-timeout==3.0.1
attrs==20.3.0
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
deepface==0.0.49
filelock==3.0.12
firebase-admin==4.5.3
Flask==1.1.2
flatbuffers==1.12
gast==0.3.3
gcloud==0.17.0
gdown==3.12.2
google-auth==1.27.1
google-auth-oauthlib==0.4.3
google-pasta==0.2.0
googleapis-common-protos==1.53.0
grpcio==1.32.0
h5py==2.10.0
httplib2==0.19.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
jws==0.1.3
Keras==2.4.3
Keras-Preprocessing==1.1.2
Markdown==3.3.4
MarkupSafe==1.1.1
mtcnn==0.1.0
multidict==5.1.0
numpy==1.19.5
oauth2client==3.0.0
oauthlib==3.1.0
opencv-python==4.5.1.48
opt-einsum==3.3.0
pandas==1.2.2
Pillow==8.1.2
protobuf==3.15.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.4.3
pyparsing==2.4.7
Pyrebase==3.0.27
PySocks==1.7.1
python-dateutil==2.8.1
python-jwt==2.0.1
pytz==2021.1
PyYAML==5.4.1
requests==2.11.1
requests-oauthlib==1.3.0
requests-toolbelt==0.7.0
rsa==4.7.2
scipy==1.6.1
six==1.15.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
tqdm==4.59.0
typing-extensions==3.7.4.3
urllib3==1.26.3
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.6.3
